use correct format for sending via mqtt

+++ src/io/SensorControl.cpp
@@ -38,6 +38,6 @@ double SensorControl::getTemperature() {
   return temperature;
 }
 
-uint16_t SensorControl::getMoisture() {
+uint32_t SensorControl::getMoisture() {
   return analogRead(MOISTURE_SENSOR_PIN);
 }
